feat: moved common, shared and core folders
