+++ lib/common/utils/analytics/base/analytic_event.dart
@@ -1,0 +1,6 @@
+import 'package:analytics/analytics.dart';
+import 'package:flutter_template/common/utils/analytics/base/has_id.dart';
+import 'package:flutter_template/common/utils/analytics/base/has_map_params.dart';
+
+/// Analytics event.
+abstract class AnalyticEvent implements AnalyticAction, HasId, HasMapParams {}
