+++ lib/common/utils/analytics/base/has_id.dart
