+++ lib/common/utils/analytics/base/has_map_params.dart
