+++ lib/config/environment/environment.dart
@@ -4,7 +4,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_template/config/app_config.dart';
 import 'package:flutter_template/config/environment/build_types.dart';
-import 'package:flutter_template/features/common/service/log_history/log_history_service_impl.dart';
+import 'package:flutter_template/common/service/log_history/log_history_service_impl.dart';
 import 'package:flutter_template/persistence/storage/config_storage/config_storage.dart';
 import 'package:flutter_template/util/log_history.dart';
 import 'package:logger/logger.dart';
