+++ lib/features/app/app.dart
@@ -4,8 +4,8 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_template/assets/themes/theme_data.dart';
 import 'package:flutter_template/config/environment/environment.dart';
 import 'package:flutter_template/features/app/di/app_scope.dart';
-import 'package:flutter_template/features/common/service/theme/theme_service.dart';
-import 'package:flutter_template/features/common/widgets/di_scope/di_scope.dart';
+import 'package:flutter_template/common/service/theme/theme_service.dart';
+import 'package:flutter_template/common/widgets/di_scope/di_scope.dart';
 import 'package:flutter_template/persistence/storage/config_storage/config_storage_impl.dart';
 
 /// App widget.
