+++ lib/features/app/di/app_scope.dart
@@ -5,14 +5,14 @@ import 'package:dio/io.dart';
 import 'package:elementary/elementary.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_template/config/environment/environment.dart';
-import 'package:flutter_template/features/common/service/theme/theme_service.dart';
-import 'package:flutter_template/features/common/service/theme/theme_service_impl.dart';
-import 'package:flutter_template/features/common/utils/analytics/amplitude/amplitude_analytic_tracker.dart';
-import 'package:flutter_template/features/common/utils/analytics/firebase/firebase_analytic_tracker.dart';
-import 'package:flutter_template/features/common/utils/analytics/mock/mock_amplitude_analytics.dart';
-import 'package:flutter_template/features/common/utils/analytics/mock/mock_firebase_analytics.dart';
-import 'package:flutter_template/features/common/utils/analytics/service/analytics_service.dart';
-import 'package:flutter_template/features/common/utils/analytics/service/analytics_service_impl.dart';
+import 'package:flutter_template/common/service/theme/theme_service.dart';
+import 'package:flutter_template/common/service/theme/theme_service_impl.dart';
+import 'package:flutter_template/common/utils/analytics/amplitude/amplitude_analytic_tracker.dart';
+import 'package:flutter_template/common/utils/analytics/firebase/firebase_analytic_tracker.dart';
+import 'package:flutter_template/common/utils/analytics/mock/mock_amplitude_analytics.dart';
+import 'package:flutter_template/common/utils/analytics/mock/mock_firebase_analytics.dart';
+import 'package:flutter_template/common/utils/analytics/service/analytics_service.dart';
+import 'package:flutter_template/common/utils/analytics/service/analytics_service_impl.dart';
 import 'package:flutter_template/features/navigation/service/router.dart';
 import 'package:flutter_template/persistence/storage/theme_storage/theme_storage.dart';
 import 'package:flutter_template/persistence/storage/theme_storage/theme_storage_impl.dart';
