+++ lib/features/dash/screen/dash_screen_wm.dart
@@ -2,9 +2,9 @@ import 'package:elementary/elementary.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_template/features/app/di/app_scope.dart';
-import 'package:flutter_template/features/common/mixin/theme_mixin.dart';
-import 'package:flutter_template/features/common/utils/analytics/event/common/track_analytics_example.dart';
-import 'package:flutter_template/features/common/utils/analytics/service/analytics_service.dart';
+import 'package:flutter_template/common/mixin/theme_mixin.dart';
+import 'package:flutter_template/common/utils/analytics/event/common/track_analytics_example.dart';
+import 'package:flutter_template/common/utils/analytics/service/analytics_service.dart';
 import 'package:flutter_template/features/dash/screen/dash_screen.dart';
 import 'package:flutter_template/features/dash/screen/dash_screen_model.dart';
 import 'package:flutter_template/l10n/app_localizations_x.dart';
