+++ lib/features/debug/screens/debug_screen/debug_screen_model.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_template/config/app_config.dart';
 import 'package:flutter_template/config/environment/environment.dart';
 import 'package:flutter_template/config/urls.dart';
-import 'package:flutter_template/features/common/service/theme/theme_service.dart';
+import 'package:flutter_template/common/service/theme/theme_service.dart';
 import 'package:flutter_template/features/debug/screens/debug_screen/debug_screen.dart';
 import 'package:flutter_template/features/debug/screens/debug_screen/debug_screen_widget_model.dart';
 import 'package:flutter_template/persistence/storage/config_storage/config_storage.dart';
