+++ lib/features/info/screen/info_screen_wm.dart
@@ -1,6 +1,6 @@
 import 'package:elementary/elementary.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_template/features/common/mixin/theme_mixin.dart';
+import 'package:flutter_template/common/mixin/theme_mixin.dart';
 import 'package:flutter_template/features/info/screen/info_screen.dart';
 import 'package:flutter_template/features/info/screen/info_screen_model.dart';
